OID MB;Gfdshif
sdf

+++ src/Buffer.java
@@ -1,6 +1,6 @@
 public class Buffer {
-    public double x1, x2, v1, v2, a, b, c;
-    public Buffer (double aa, double bb, double cc) {
+    public int x1, x2, v1, v2, a, b, c;
+    public Buffer (int aa, int bb, int cc) {
         x1 = 0; x2 = 0; v1 = 0; v2 = 0;
         a = aa;
         b = bb;
